+++ core/src/main/scala/org/gerweck/scala/util/ObjectTree.scala
@@ -1,5 +1,6 @@
 package org.gerweck.scala.util
 
+import scala.collection.compat._
 import scala.reflect.runtime.universe._
 
 import org.log4s._
@@ -26,7 +27,7 @@ object ObjectTree {
     var hitDepthLimit: Boolean = false
     def ind(s: String) = s.linesWithSeparators.map(" " * indent + _).mkString
     def singleLine(s: String) = {
-      s.size < singleLineSizeLimit && s.lines.drop(1).isEmpty
+      s.size < singleLineSizeLimit && s.linesIterator.drop(1).isEmpty
     }
     def smartShow(any: Any, currentDepth: Int): String = {
       if (currentDepth > maxDepth) {
@@ -53,15 +54,15 @@ object ObjectTree {
           val shownValue = smartShow(v, depth)
           ind(showKV(showKey(k), smartShow(v, depth), true))
         case a: TraversableOnce[_] =>
-          a.toIterator
-            .map(smartShow(_, depth))
-            .mkString("\n")
+          a.iterator
+           .map(smartShow(_, depth))
+           .mkString("\n")
         case p: Product if p.productArity == 1 && singleLine(smartShow(p.productElement(0), depth)) =>
           p.toString
         case a: Product if a.productArity > 0 =>
           val aMirror = currentMirror.reflect(a)
           val fieldStrings =
-            aMirror.symbol.typeSignature.members.toStream
+            aMirror.symbol.typeSignature.members.to(Stream)
               .collect { case a: TermSymbol => a }
               .filter(a => a.isParamAccessor)
               .filterNot(a => a.isMethod || a.isModule)