+++ core/src/main/scala/org/gerweck/scala/util/Resources.scala
@@ -1,6 +1,6 @@
 package org.gerweck.scala.util
 
-import scala.collection.JavaConverters._
+import scala.jdk.CollectionConverters._
 
 import java.io.File
 import java.net.URL
@@ -68,7 +68,7 @@ object Resources {
       }
     }
 
-    (Map.empty[String,URL] /: maps) (_ ++ _)
+    maps.foldLeft(Map.empty[String,URL]) (_ ++ _)
   }
 
   def forPackage(name: String): Map[String, URL] = forPackage(Package.getPackage(name))